(NETDEV-26) Add network_dns type
Without this patch there is no type to manage the DNS configuration of a
target device.  This patch addresses the problem by adding the
network_dns type with a namevar of `domain` and array properties of
`search` and `servers`.  The type is ensurable because DNS may not be
configured on the target system.