modifies newtork_dns to remove ensurable
This proposed update removes the ensureable setting from network_dns since it
does not make sense.  The network_dns type configures global settings for
dns domain, search and servers. There are no unique instances of type
network_dns therefore it does not make sense for the type to be ensurable.

For specifying ensurable entries in the search and servers arrays the
name_server and search_domain types should be used instead.